components/livekit/examples/voice_agent/agent/pyproject.toml
[project]
name = "esp-example-agent"
version = "0.1.0"
description = "Example agent for ESP32 voice agent example."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[openai]~=1.2.14",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1"
]